2010-06-20 19:16 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * src/vm/macro.c   * src/vm/estack.c   * src/vm/dynsym.c   * src/vm/hvm.c   * src/vm/set.c   * src/rdd/wafunc.c   * src/compiler/hbmain.c   * src/compiler/cmdcheck.c   * src/compiler/genc.c   * src/compiler/gencc.c   * src/compiler/harbour.yyc   * src/compiler/harbour.y   * include/hbexprb.c     + Added strlen() casts to pacify msvc64 warnings
  * src/vm/cmdarg.c
    ! Readded one '( void * )' cast. (I've only tested C++ mode
      on Windows, this fell into *nix code). Thanks for checking it.

  * src/rtl/hbgtcore.c
    + HB_SIZE -> HB_UINT, plus added necessary casts. Signaled by warnings

  * contrib/hbmisc/spd.c
    + HB_SIZE -> HB_UINT, plus added necessary casts. Signaled by warnings
    ; In case of this function it'd be probably better to make it fully
      HB_SIZE compatible, instead of stepping back like this.

  * src/rdd/sdf1.c
  * src/rdd/delim1.c
    + Added casts to pacify some msvc64 warnings

  * src/rdd/hsx/hsx.c
    ! HB_SIZE -> HB_LONG. Resolves msvc64 warnings.

+++ harbour/include/hbexprb.c
@@ -5443,7 +5443,7 @@ static HB_EXPR_PTR hb_compExprReduceAliasString( HB_EXPR_PTR pExpr, HB_EXPR_PTR 
 
 static HB_BOOL hb_compExprIsMemvarAlias( const char *szAlias )
 {
-   int iLen = strlen( szAlias );
+   int iLen = ( int ) strlen( szAlias );
    /* @M-> @MEMVAR-> or @MEMVA-> or @MEMV-> */
    return ( iLen == 1 || ( iLen >= 4 && iLen <= 6 ) ) &&
           memcmp( szAlias, "MEMVAR", iLen ) == 0;
