+++ harbour/src/vm/estack.c
@@ -1220,7 +1220,7 @@ void hb_stackDispCall( void )
 
    while( hb_procinfo( iLevel++, buffer, &uiLine, file ) )
    {
-      int l = strlen( buffer );
+      int l = ( int ) strlen( buffer );
       hb_snprintf( buffer + l, sizeof( buffer ) - l, "(%hu)%s%s", uiLine, *file ? HB_I_(" in ") : "", file );
 
       hb_conOutErr( "Called from ", 0 );
